+++ TVRename/Utility/Helper/RegistryHelper.cs
